Rework how user primary keys are stored in the metadata, and add some tests
EVIL HACK ALERT.  This depends on the fact that if you have a multi-
column UNIQUE constraint on a table, it's ignored for rows which contain
NULL in one or more columns.  So on the columns_map table we create such
a constraint on (dataset_system_id, lifecycle_version, is_primary_key)
and then make is_primary_key (which has type "unit null") not-null on
the PK column.