Make *Info objects JSONable
Not too happy with this yet.  In particular, I'm wondering if a ColumnInfo
codec should require a VersionInfo to construct (and same for the
VersionInfo/DatasetInfo relation).  As it is now, serializing a schema
with n columns will produce n identical copies of version and datset info.

But the only time we serialize more than one is when a dataset is
truncated, which should be rare, and it's not THAT much space...