Stream the generated insert-CSV a little better.
I'm not _quite_ sure why but this seems to perform better than simply
producing a complete CSV for the insert-batch in-memory and then sending
it off to Postgres.  Now we generate it in 100k chunks and (I believe!)
make the next while that's processing.  Certainly it takes ~30s off the
Crimes import on my workstation.

The magic "100000, 10" for the chunk size + total buffered chunks can
probably be tweaked.