Rework dropping (working) copies
Before it was a strange mash of dropping snapshots & dropping working
copies.

I'm still not too happy with it.  It would be better to completely separate
these two concepts (since only "drop working copy" is very special;
"drop snapshot" sort of is since it affects a copy that is not the current
copy, but only "drop working copy" moves back in time)