Extensive but not major refactoring preparing for the concept of a "data context"
A "data context" is the complete notion of "the truthstore", including
types, values, limits, operations, the actual physical kind of database,
etc.  This is the glue that makes all the parts work together.

The biggest change here is passing in a function that stands in for
conn.getCopyAPI.copyIn(sql, csv) instead of passing in a function that
just gets the copy API.