Stream reports from mutation scripts
This entailed a mutation-script-report format change.  Whereas before
it came back in an list-of-objects-of-objects that looked like
  [
    { // each element is the results of one row_data command
      "inserted": { "1":{}, "2": {}, ...}
      "updated": ...
       ...
    },
    ...
  ]
now it comes back as a list of lists with one (inner) element for each
job, in the order of each job's submission.  (it's still one outer
element for each row_data command)
  [
     [ {"typ":"inserted","id":1,"version":"rv-xxxx.xxxx.xxxx"},
       {"typ":"inserted", ... },
       ...
     ],
     ....
   ]
The details of the inner elements will probably change to reduce
redundancy a bit.