Add the concept of secondary groups.
A secondary group is a set of secondary instances.  We are taking a shortcut and
overloading the existing /secondary-manifest/<secondary>/<dataset> call.  The call
can now take either a secondary name as before, or a string "_foo_" where foo is a
secondary group, or the magic string "_DEFAULT_" which will publish to all groups
in the default set of groups, as defined in the config.

Each group supports a numReplicas setting to say how many different
instances within the group the dataset should be published to.

At this point, we do not store any information about what groups a dataset is published
to, we only know what secondary instances it ended up on.