Add support for rewriting function calls directly to column refs.
This enables rewriting a function call from the query directly to
a column ref in the rollup, removing the function call entirely.
For example, if the rollup has date_trunc_ym(date) and the query
uses the same function, it just gets converted to refer to the
column in the rollup directly.