Add support for rewriting date queries to use rollups with date_trunc_xxx columns.
We do that by looking to see if the literal value passed in is the same as the output from a
date_trunc_xxx call, and if so substituting it where "column > literal" or "column <= literal".
The forms "literal < column" and "literal >= column" are easy to add but not implemented.

See the scaladoc for the rewriteDateTruncGteLt method in QueryRewriter for more details.