CORE-7254: Incorrectly rewriting queries without a GROUP BY to use a grouped rollup.
In the case where the query didn't have a GROUP BY on it, but the
rollup did, we could incorrectly rewrite the query to use the rollup
resulting in incorrect results if all the columns in the query are
in a rollup.

We now will only do the rewrite if all the columns are aggregates.
The Selection rewrite will ensure the aggregates can be mapped
appropriately to the rollup's Selection by either being self
aggregatable or falling into some other specific special aggregation
case.