Add basic support for selecting the best rollup when doing query rewrite.
This is essentially a simple "rule based" optimizer that looks at the structure of the
rollup to assign it a score based on how expensive we think it might be to query.
We do not consider the particulars of the query, just the rollup.

Please see the docs in com.socrata.querycoordinator.RollupScorer for an explaination
of how it functions.