EN-1444: Increase the timeout for query coordinator from 2m to 10m.
We have various slow queries. Right now query coordinator times
them out after 2 minutes, but they keep running on the backend.
This isn't a good state because they will then keep getting run but
will never get cached or returned to the user.

Increasing this to 10 minutes allows us to successfully run and
cache a wider variety of queries. Note I think soda-fountain will
still timeout after 5 minutes but we can address that if necessary.

Also, add support for setting it in the docker config file in case
we need to further tweak it.