EN-13225:Commit before updating secondary-manifest
We want to commit before updating the `latest_secondary_data_version`
to end the current transaction during replay to try to avoid deadlocks
when data-coordinator is also updating the `latest_data_version` for
the same row and is writing to the log tables that the secondary-watcher
was also trying to read from.

Deadlocks at the end of replay cause the secondary-watcher to go into
a resync which causes the dataset to not be available for read.