EN-15144: Add column for dataset resource_name
We want to start storing dataset resource_name in truth
so the secondary-watcher can include the external dataset
ids in messages in AMQ reporting replication complete.

These replication messages are so we can wait to calculate
column stats after replication is complete.

Also, update migrations script to use sbt instead of first
assembling a new jar (because that is slow).