Explicitly pin Scala version.
This is a good thing to do anyway, but the specific reason for doing
it is it looks like different versions are fighting resulting in
strange "java.lang.ClassNotFoundException: scala.Some" errors.

I briefly looked at upgrading past 2.10, but we are still blocked by the
sbt-scalabuff plugin, which we would need to either fork and update or
replace.