EN-19241: Add by_system_id option to row_data cmd
Add by_system_id option to row_data command in the mutation script.

The option only changes the behavior of upserts and deletes for
datasets with user primary keys set in the following ways:

 - updates may be made to rows using the system id column.
 - if a user primary key column value and a system id colum
   value is given, the user primary key is prefered.
 - all deletes are assumed to be made using the system id column.

This option is to support the FeedbackSecondary library so that
a feedback secondary does not need to track the user primary key
column and not have to make guesses about if the primary key
column has changed based on errors from data-coordinator.