EN-5487: Support delete on S-M for full stores
Support delete on /secondary-manifest for full secondary stores.

Add configuration for specifying instances in a secondary group
that are full. With this change, we will not remove secondary
instances in a group from data-coordinator's config that are
full, but instead add them to the set of full instances for
the group in question.

This fixes delete not working on /secondary-manifest for full stores
that have been removed from data-coordinator's config by not removing
them from data-coordinator's config to start with.

This will also support data-coordinator's future collocation logic
to be able to know about all instances in a secondary group, but
still know about which stores have been backlisted as too full
to move datasets to.