EN-17021: secondary_metrics table exists
Database miratgion for secondary_metrics table.

Table to store the "total size" of a dataset on a secondary store
to support a more useful cost function for collocation alogrithm.

Note: with the fkey to the secondary_manifest this database migration
may take some extra care to be run in production.