EN-21686: Add size to cost vector in collocation
Add total size and maximun (individual) move size to
cost vector for collocation in addition to the number
of moves. The ordering for cost is configurable with
weights.

For the pg-secondary implementation, the size here is
the total size of t-tables with indices of datasets
on the secondary-stores.

Includes what I think is a reasonable starting point
for weights in `reference.conf`.