EN-17026: Collocation API can reject collocation
Add limits to collocation cost passed in the collocation
request such that data-coordinator can reject collocation
when the cost exceeds those limits.

Add a new `"limits"` field to the collocation request body
pointing to a Cost object.

Also, continue rename of initiateCollocation -> executeCollocation
in Unit test.

Note: this is a breaking change! Since the new limits field
is required.