EN-24804: Fix collocation manifest recursive query
Because of the UNION ALL in the recursive query on the collocation
manifest causes the rows at each depth to increase factorially in
number (a problem for datasets with a large branching factor), the
query is far too slow -- and worse -- creates very large temp tables.

Replace the query with a breadth first search of the collocation
manifest implemented in scala.