Allow gating updates by expected dataset version
We already send back the new version on every mutation in a header, so
it is possible to do multiple writes while guaranteeing that no one
has touched the dataset in between.

This also presents the data version on exports, so it will be possible
to do an atomic read-modify-update cycle with dataset-level granularity.

This is all in support of doing incremental updates to datasets
externally.