EN-41342: Consolidate multiple row-update-only data versions into one big data version (#309)
In particular you can get a ton of such versions when the feedback
secondaries discover a big change.  This has the potential to make things
MUCH faster on large updates to large datasets even, because it can mean
the feedback secondary has half a hope of triggering the "make an indexless
copy" logic.