Prune secondary sets when collocating
Currently collocation is sort of half-aware of secondary groups, but
it is not fully aware -- in particular, when it asks another DC where
the secondaries for a remote dataset live, it does not trim the response
it gets down to the secondary-set that it's interested in.

This makes it do so.