EN-52670: Add column rollup_map.raw_soql (#351)
The field rollup_map.soql is a transformed soql with internal column ids which is not good for human readability. In theory, we have code to convert this soql back and forth. But it is a difficult process. Bugs can also get in the way. It would be nice to save a copy of the original soql given by the user.

The field is not used in any computation.  It is only for developer/operation convenience.