Allow ourselves to introspect query plans for specific upsert operations
This lets us poke an endpoint to say "hey, the next time this specific
thread starts doing a non-INSERT thing on the database as part of an
upsert operation, log the output of EXPLAINing that query".  INSERTs
are done via COPY which it doesn't really super make sense to EXPLAIN,
but DELETEs, UPDATEs, and the SELECTs to find existing rows all count.