Moving datasets between truths, part 1
With this code, we can copy datasets between truths.  What you'll
end up with is a duplicate dataset in the target truth which lives
on _no_ secondaries.  What yet needs to happen:
  * Be sure replication is fully up-to-date before even starting this
  * Tell the secondaries "hey, dataset alpha.1234 is now bravo.4321"
  * add secondary-manifest records appropriately on the new secondary

Step 2 of that is the tricky one (it's easy for PG secondaries, not
nearly so easy for spandex/archival)