Cross-truth dataset mover (#377)
* Moving datasets between truths, part 1

With this code, we can copy datasets between truths.  What you'll
end up with is a duplicate dataset in the target truth which lives
on _no_ secondaries.  What yet needs to happen:
  * Be sure replication is fully up-to-date before even starting this
  * Tell the secondaries "hey, dataset alpha.1234 is now bravo.4321"
  * add secondary-manifest records appropriately on the new secondary

Step 2 of that is the tricky one (it's easy for PG secondaries, not
nearly so easy for spandex/archival)

* Moving datasets between truths, part 2

* Handle datasets that live only in geocoding + pg secondaries

* Fix misc bugs

* Add a dry-run mode

* Sanity-check that from/to are c3p0-less

* Rearrange configuration

Now there's just a set of truths, and you say "move alpha.12345 into
bravo" or whatever on the command line.

* Make mover configuration a little nicer

* Nicer UI

* Copy the audit table too

They're not particularly useful, but it's easy to do so might as well.

* Fixes

* Add a README to explain the config file format