EN-78533: Deleting datasets didn't use the `pending_drop` machinery
This is all ancient code, so old it predates `pending_drop`, which was
added to support moving datasets around secondaries.  When a dataset
was deleted, its `latest_data_version` and `went_out_of_sync_at` were
both moved into the future - the latter so that the periodic table
cleanup wouldn't eat the `secondary_manifest` row for the deleted dataset
too quickly.  Then the secondary watcher would say "this dataset needs
updating because its data versions don't match, oh the dataset doesn't
exist, clean it up."

BUT!  A long time ago we added `pending_drop` to allow removing a dataset
from a secondary while it was still alive, and the secondary-watcher
seeks out entries where the data versions don't match _or_ `pending_drop`
is true - in the latter case it removes the `secondary_manifest` row
itself.  So we can use that instead of the timestamp hack, which is good
because apparently something is resetting `went_out_of_sync_at` back to
a "real" value causing prompt cleanup of the `secondary_manifest row`.